+++ sample/src/main/assets/app.properties
@@ -6,4 +6,3 @@ system.toaster.centered=false
 encrypt.shared.preferences=false
 allow.offline.login.with.invalid.token=false
 ignore.location.deletion = false
-monitor.height=true
